requirements.txt
azure-identity
azure-storage-blob==12.18.3
azure-search-documents==11.6.0b1
langchain==0.1.2
langchain-core==0.1.17
langchain-community==0.0.16
langchain-openai==0.0.6
python-dotenv
unstructured[pdf,xlsx,docx,pptx]==0.12.6